Move text_utils.is_acronym => extract.acronyms.is_acronym
